Refactor import paths and adjust authentication directory
