+++ lib/views/auth/verified_account_client.dart
